Very minor editorial changes, addition to techniques
Do we need (to refer to) a definition of 'essential' - i.e. does it mean "can be done in no other way" or also "would be inconvenient / cumbersome to do in another way"? Authors may claim the way they things do is essential so they "can't be asked" to think of more accessible ways. But perhaps there is always a grey area here which no amount of definition can remove.

# Techniques: I do not understand "confirmation dialogue re-appears each time focus leaves the signature panel." Why? When focus leaves the sdignature panel, AT gestures work again? Mighr be a violation of 3.2.1 On Focus?